src/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
decouple==0.0.7
dj-rest-auth==7.0.1
Django==5.1.7
django-allauth==65.4.1
django-configurations==2.5.1
django-phonenumber-field==8.0.0
django-rest-auth==0.9.5
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
idna==3.10
phonenumbers==9.0.0
pillow==11.1.0
pycparser==2.22
PyJWT==2.9.0
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
urllib3==2.3.0
djangorestframework
django-cors-headers
django-filter
djoser
django-taggit
channels
django-celery-beat
django-celery-results
django-debug-toolbar
boto3
django-storages
django-crispy-forms
crispy-tailwind
